+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Exception in `add()` for names starting with underscore
+
 ### Changed
 - Refactored `remove()` and `__delitem__()` to use shared `_del_property()` helper method
-- Hint in get() method
+- Proper hint in get() method
 
 ## [1.0.1] - 2025-11-18